# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Query Assistant API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@yourcompany.com"
        },
        "version": "1.0"
    },
    "host": "dev.smartbet.live",
    "basePath": "/query-assistant/api/v1",
    "paths": {
...
        "/query/execute": {
            "post": {
                "description": "Convert a natural language prompt to SQL and execute it against ClickHouse with site_id filtering for multi-tenancy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "Execute a natural language query with site filtering",
                "parameters": [
                    {
                        "description": "Query request with required site_id for multi-tenant filtering",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful query execution with results filtered by site_id",
                        "schema": {
                            "$ref": "#/definitions/models.QueryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request - missing site_id or invalid format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Query timeout",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/query/generate": {
            "post": {
                "description": "Generate a SQL query from natural language prompt without executing it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "Generate SQL query",
                "parameters": [
                    {
                        "description": "Generation request with site_id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QueryGenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated SQL query with metadata",
                        "schema": {
                            "$ref": "#/definitions/models.QueryGenerateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/query/history": {
            "get": {
                "description": "Retrieve the history of previously executed queries with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "Get query history",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by site ID",
                        "name": "site_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of historical queries",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.QueryHistory"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/query/validate": {
            "post": {
                "description": "Validate a SQL query for syntax and security without executing it, with optimization suggestions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "query"
                ],
                "summary": "Validate a SQL query",
                "parameters": [
                    {
                        "description": "Query to validate with site_id",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QueryValidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Validation result with warnings and optimization suggestions",
                        "schema": {
                            "$ref": "#/definitions/models.QueryValidationResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ColumnMetadata": {
            "type": "object",
            "properties": {
                "database_type": {
                    "type": "string",
                    "example": "UInt64"
                },
                "name": {
                    "type": "string",
                    "example": "user_id"
                },
                "nullable": {
                    "type": "boolean",
                    "example": false
                },
                "position": {
                    "type": "integer",
                    "example": 0
                },
                "sample_value": {
                    "type": "string",
                    "example": "12345"
                },
                "type": {
                    "type": "string",
                    "example": "UInt64"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "object"
                },
                "error": {
                    "type": "string",
                    "example": "Query validation failed"
                },
                "message": {
                    "type": "string",
                    "example": "Query contains forbidden operation: DROP"
                },
                "query_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
...
        "models.PaginationInfo": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "has_previous": {
                    "type": "boolean",
                    "example": false
                },
                "next_page": {
                    "type": "integer",
                    "example": 2
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 100
                },
                "previous_page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                },
                "total_rows": {
                    "type": "integer",
                    "example": 1000
                }
            }
        },
        "models.QueryGenerateRequest": {
            "type": "object",
            "required": [
                "prompt",
                "site_id"
            ],
            "properties": {
                "prompt": {
                    "type": "string",
                    "example": "Show GGR for last month"
                },
                "site_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 123
                }
            }
        },
        "models.QueryGenerateResponse": {
            "type": "object",
            "properties": {
                "complexity": {
                    "type": "string",
                    "example": "medium"
                },
                "generated_sql": {
                    "type": "string",
                    "example": "SELECT SUM(bet_amount) - SUM(win_amount) as GGR FROM transactions WHERE site_id = 123 AND created_at \u003e= today() - 30 LIMIT 1"
                },
                "prompt": {
                    "type": "string",
                    "example": "Show GGR for last month"
                },
                "site_id": {
                    "type": "integer",
                    "example": 123
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "models.QueryHistory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "error": {
                    "type": "string"
                },
                "execution_time": {
                    "type": "number",
                    "example": 0.156
                },
                "generated_sql": {
                    "type": "string",
                    "example": "SELECT month, SUM(sales) FROM sales_table WHERE site_id = 123 GROUP BY month"
                },
                "prompt": {
                    "type": "string",
                    "example": "Show total sales by month"
                },
                "query_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "row_count": {
                    "type": "integer",
                    "example": 12
                },
                "site_id": {
                    "type": "integer",
                    "example": 123
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "models.QueryRequest": {
            "type": "object",
            "required": [
                "site_id"
            ],
            "properties": {
                "context": {
                    "type": "object"
                },
                "include_schema": {
                    "description": "Response options",
                    "type": "boolean",
                    "example": true
                },
                "include_stats": {
                    "type": "boolean",
                    "example": true
                },
                "page": {
                    "description": "Pagination options",
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 1,
                    "example": 100
                },
                "prompt": {
                    "type": "string",
                    "example": "Show me the top 10 users by total purchase amount in the last 30 days"
                },
                "site_id": {
                    "description": "Required site_id parameter (numeric)",
                    "type": "integer",
                    "minimum": 1,
                    "example": 123
                },
                "timeout": {
                    "type": "integer",
                    "maximum": 300,
                    "minimum": 1,
                    "example": 30
                }
            }
        },
        "models.QueryResponse": {
            "type": "object",
            "properties": {
                "cache_key": {
                    "type": "string"
                },
                "columns": {
                    "description": "New fields for column metadata and pagination",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ColumnMetadata"
                    }
                },
                "execution_time": {
                    "type": "number",
                    "example": 0.234
                },
                "generated_sql": {
                    "type": "string",
                    "example": "SELECT user_id, SUM(amount) as total FROM purchases WHERE site_id = 123 AND created_at \u003e= today() - 30 GROUP BY user_id ORDER BY total DESC LIMIT 10"
                },
                "pagination": {
                    "$ref": "#/definitions/models.PaginationInfo"
                },
                "prompt": {
                    "type": "string",
                    "example": "Show me the top 10 users by total purchase amount"
                },
                "query_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "row_count": {
                    "description": "Legacy fields (kept for backward compatibility)",
                    "type": "integer",
                    "example": 10
                },
                "statistics": {
                    "$ref": "#/definitions/models.QueryStatistics"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "warnings": {
                    "description": "Additional metadata",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['Query processed large amount of data']"
                    ]
                }
            }
        },
        "models.QueryStatistics": {
            "type": "object",
            "properties": {
                "bytes_processed": {
                    "type": "integer",
                    "example": 1048576
                },
                "cache_hit": {
                    "type": "boolean",
                    "example": false
                },
                "execution_time_ms": {
                    "type": "number",
                    "example": 234.5
                },
                "partitions_accessed": {
                    "type": "integer",
                    "example": 3
                },
                "rows_returned": {
                    "type": "integer",
                    "example": 100
                },
                "total_rows": {
                    "type": "integer",
                    "example": 50000
                }
            }
        },
        "models.QueryValidateRequest": {
            "type": "object",
            "required": [
                "query",
                "site_id"
            ],
            "properties": {
                "query": {
                    "type": "string",
                    "example": "SELECT * FROM users WHERE site_id = 123 LIMIT 10"
                },
                "site_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 123
                }
            }
        },
        "models.QueryValidationResult": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "estimated_cost": {
                    "type": "string",
                    "example": "low"
                },
                "estimated_rows": {
                    "description": "New fields",
                    "type": "integer",
                    "example": 10000
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "optimizations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "valid": {
                    "type": "boolean"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
...
}